feat: set default observedGeneration to -1 on HelmReleases
This resolves an issue with kustomize-controller marking a
Kustomization as healthy even when the helm-controller hasn't even
looked at the HelmRelease targeted by the Kustomization's
healthChecks, yet. Setting `observedGeneration` to -1 by default will
cause kstatus to report a status of `InProgress` instead of `Ready`.

see https://github.com/fluxcd/kustomize-controller/issues/387 for
details on the issues this is solving.

Signed-off-by: Max Jonas Werner <mail@makk.es>